+++ tests/saturation/test_distances.py
@@ -7,7 +7,7 @@ from saturation.distances import Distances
 
 
 def _create_crater(id: int, x: float, y: float, radius: float) -> Crater:
-    return Crater(id=id, x=np.float32(x), y=np.float32(y), radius=np.float32(radius))
+    return Crater(id=id, x=np.float64(x), y=np.float64(y), radius=np.float64(radius))
 
 
 def assert_float_equal(first: float, second: float, *, percentage: float = None, absolute: float = None):
